dev/counter: introduce primitive counter dev
In this commit we implement the first version of primitive counting
device. This is (for now) placed to a hardcoded address in the memory
map of our VM (0x5000) and provides 2 registers: one at 0x5000 -
readable. Second one at 0x5004 - writable. Every write to the second
register advances counter readable in the first register.

Also add a basic improvement to the guest assembly code to test this
functionality. Note that the guest code will only print the counter
properly if its value is < 10.

Signed-off-by: Jakub Acs <acsjakub@gmail.com>